+++ CHANGELOG.md
@@ -35,3 +35,4 @@
 ## - 0.4.0
 * Added the ability to manipulate widget settings by setting them in the flow
 * Added the ability to add custom HTML/widgets to the horizontal/vertical stack
+* Fixed a bug where it was possible to add a hor/vert stack to itself causing a circular dependency loop crashing the dashbored